Convert LLVM native functions to TypeScript
